refactor: extract PeersConfig
