+++ crates/net/network-types/Cargo.toml
@@ -1,0 +1,31 @@
+[package]
+name = "reth-network-types"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+description = "Commonly used network types"
+
+[lints]
+workspace = true
+
+[dependencies]
+# reth
+reth-network-api.workspace = true
+reth-network-peers.workspace = true
+reth-net-banlist.workspace = true
+
+# io
+serde = { workspace = true, optional = true }
+humantime-serde = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+
+# misc 
+tracing.workspace = true
+
+[features]
+default = ["serde"]
+serde = ["dep:serde", "dep:humantime-serde", "dep:serde_json"]
+test-utils = []
