+++ crates/net/network-types/src/peers/mod.rs
@@ -1,0 +1,5 @@
+pub mod reputation;
+pub use reputation::ReputationChangeWeights;
+
+pub mod config;
+pub use config::{ConnectionsConfig, PeersConfig};
